nptl: Fix racy pipe closing in tst-cancel{20,21}
The tst-cancel20 open two pipes and creates a thread which blocks which
blocks reading the first pipe.  It then issues a signal to activate a
signal handler which block s reading the second pipe.  Finally the
cancellation cleanup-up handlers are tested by first closing the
all the pipes ends and issuing a pthread_cancel. Tst-cancel21 have
a similar behavior, but use an extra fork after the test itself.

The race condition occurs if the cancellation handling acts after the
pipe close: in this case read will return EOF (indicating side-effects)
and thus the cancellation must not act.  However current GLIBC
cancellation behavior acts regardless the syscalls returns with
sid-effects.

This patch adjust the test by moving the pipe closing after the cancellation
handling.  This avoid spurious cancellation if the case of race described.

Checked on x86_64 and i386.

	* nptl/tst-cancel20.c (do_one_test): Move the pipe closing after
	pthread_join.
	* nptl/tst-cancel21.c (tf): Likewise.