Move ASM_SIZE_DIRECTIVE to generic sysdep.h
This patch moves the ASM_SIZE_DIRECTIVE to generic sysdep.h since
current GLIBC only supports ELF and the definition is duplicated
among all the architectures that uses it.

No code generation change is expected from this patch.  I tested
glibc build for the affected ports and no build regressions were
found.

It fixes BZ#14414, although with a different approach (it does not
get rid of the macro, but instead move its definition to a common
place).

	* sysdeps/aarch64/sysdep.h (ASM_SIZE_DIRECTIVE): Remove macro.
	* sysdeps/i386/sysdep.h (ASM_SIZE_DIRECTIVE): Likewise.
	* sysdeps/ia64/fpu/libm-symbols.h (ASM_SIZE_DIRECTIVE): Likewise.
	* sysdeps/microblaze/sysdep.h (ASM_SIZE_DIRECTIVE): Likewise.
	* sysdeps/nios2/sysdep.h (ASM_SIZE_DIRECTIVE): Likewise.
	* sysdeps/powerpc/sysdep.h (ASM_SIZE_DIRECTIVE): Likewise.
	* sysdeps/s390/s390-32/sysdep.h (ASM_SIZE_DIRECTIVE): Likewise.
	* sysdeps/s390/s390-64/sysdep.h (ASM_SIZE_DIRECTIVE): Likewise.
	* sysdeps/sh/sysdep.h (ASM_SIZE_DIRECTIVE): Likewise.
	* sysdeps/tile/sysdep.h (ASM_SIZE_DIRECTIVE): Likewise.
	* sysdeps/unix/sysv/linux/powerpc/powerpc64/sysdep.h
	(ASM_SIZE_DIRECTIVE): Likewise.
	* sysdeps/x86_64/sysdep.h (ASM_SIZE_DIRECTIVE): Likewise.
	* sysdeps/generic/sysdep.h (ASM_SIZE_DIRECTIVE): New macro