libio: Fix seek-past-end returned size for open_{w}memstream (BZ#15298)
This patch fixes how open_{w}memstream updates the returned buffer
size on a fclose/fflush operation.  POSIX states [1]:

"After a successful fflush() or fclose() [...] the variable pointed
to by sizep shall contain the smaller of the current buffer length and
the number of bytes for open_memstream(), or the number of wide characters
for open_wmemstream(), between the beginning of the buffer and the current
file position indicator."

Current GLIBC behavior returns the seek position even if there is no previous
write operation.  To correctly report the buffer size the implementation must
track both the buffer positiob and currently bytes written.  However internal
_IO_write_ptr is updated on both write and seek operations.

This patch fixes it by adding two new internal fields to keep track of both
previous and next position after a write operation.  It allows the sync and
close callbacks to know if a write has occurred based on previous and
current _IO_write_ptr value.

Tested on x86_64.

	[BZ #15298]
	* libio/memstream.c  (_IO_FILE_memstream): Add prevwriteptr and
	seekwriteptr fields.
	(_IO_mem_seekoff): New function.
	(_IO_mem_jumps): Use _IO_mem_seekoff.
	(__open_memstream): Initialize prevwriteptr and seekwriteptr.
	(_IO_mem_sync): Update sizeloc based on written bytes instead of buffer
	current position.
	(_IO_mem_finish): Likewise.
	* libio/wmemstream.c  (_IO_FILE_wmemstream): Add prevwriteptr and
	seekwriteptr fields.
	(_IO_wmem_seekoff): New function.
	(_IO_wmem_jumps): Use _IO_mem_seekoff.
	(__open_wmemstream): Initialize prevwriteptr and seekwriteptr.
	(_IO_mem_wsync): Update sizeloc based on written bytes instead of buffer
	current position.
	(_IO_mem_wfinish): Likewise.
	* libio/tst-memstream3.c (do_test_bz15298): New function.
	(do_test_bz18241): Check for expected size after a fseek followed by a
	fflush.
	(do_test): Call do_test_bz15298.

[1] http://pubs.opengroup.org/onlinepubs/9699919799/functions/open_memstream.html