Use gcc stpncpy chk builtin (BZ #20661)
Since r182378 GCC provides stpncpy_chk builtin support and it is
included in official GCC 4.7.4 package.  It allows GLIBC to use it
on string3.h headers if GCC 4.7 or higher is used when including
this headers.

This patch changes current fortify code to use it instead. The
fortify definition is also new guarded with __USE_GNU, similar to
stpcpy.

Check on x86_64 with GCC 5.4.

	[BZ #20661]
	* string/bits/string3 [__USE_GNU] (stpncpy): Add fortify version based
	on GCC builtin.