nptl: alpha: Fix Race conditions in pthread cancellation (BZ#12683)
This patch adds the alpha modifications required for the BZ#12683.
It basically adds the required __pthread_get_pc function

The default syscall_cancel.c should be fine for alpha and GCC 6.1
generates the following assembly for syscall_cancel.c:

 # 52 "../sysdeps/unix/sysv/linux/syscall_cancel.c" 1
        callsys # $0 $19 <= $16 $17 $18 $20 $21 $0 $19
 # 0 "" 2
$LVL8:
        .set    nomacro
$LBE3:
$LBE2:
        .loc 1 53 0
        .set    macro
 # 53 "../sysdeps/unix/sysv/linux/syscall_cancel.c" 1
        .global __syscall_cancel_arch_end
__syscall_cancel_arch_end:

Checked against a build and make check run-built-tests=no for
alpha-linux-gnu.

	* sysdeps/unix/sysv/linux/alpha/sysdep-cancel.h (__pthread_get_pc):
	New function.