nptl: tile: Fix Race conditions in pthread cancellation (BZ#12683)
This patch adds the tile modifications required for the BZ#12683.
It basically adds the required __pthread_get_pc function

The default syscall_cancel.c should be fine for tile and GCC 6.1
generates the following assembly for syscall_cancel.c:

[...]
        swint1
.LVL5:
.LBE2:
        .loc 1 53 0
        .global __syscall_cancel_arch_end
__syscall_cancel_arch_end:
[...]

Checked against a build and make check run-built-tests=no for
tile-linux-gnu.

	* sysdeps/unix/sysv/linux/tile/sysdep-cancel.h (__pthread_get_pc):
	New function.